fix: delete user 시 cookie 제거되지 않던 문제 해결
- delete user 시 clearCookie
- delete user 시 redis delKey

+++ backend/src/users/users.controller.ts
@@ -8,12 +8,17 @@ import {
   Delete,
   ParseIntPipe,
   UseGuards,
+  UnauthorizedException,
+  Redirect,
+  Res,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { ApiCreatedResponse, ApiOperation, ApiTags } from '@nestjs/swagger';
 import { User } from './entities/user.entity';
 import { AuthGuard } from '@nestjs/passport';
+import { GetCurrentUserId } from '../commons/decorators/get.current-userId.decorator';
+import { Response } from 'express';
 
 @Controller('users')
 @ApiTags('Users API')
@@ -38,7 +43,18 @@ export class UsersController {
   @ApiOperation({ summary: '유저 삭제 API', description: '유저를 삭제합니다.' })
   @ApiCreatedResponse({ description: '유저를 삭제합니다.', type: null })
   @UseGuards(AuthGuard('jwt'))
-  removeUserById(@Param('id', ParseIntPipe) id: number): Promise<void> {
+  removeUserById(
+    @Param('id', ParseIntPipe) id: number,
+    @GetCurrentUserId() currentUserId: number,
+    @Res({ passthrough: true }) res: Response,
+  ): Promise<void> {
+    if (id !== currentUserId) {
+      throw new UnauthorizedException();
+    }
+
+    res.clearCookie('AccessToken');
+    res.clearCookie('RefreshToken');
+
     return this.usersService.removeUserById(id);
   }
 }
